refactor(projects/art-jam-i): Moves svg component under src/components/p5
